feat: move pages to auth folder & add img for product
