+++ frontend/src/App.jsx
@@ -7,8 +7,8 @@ import Header from "./components/Header";
 import Footer from "./components/Footer";
 import NotFoundPage from "./pages/NotFoundPage";
 import ProductDetail from "./pages/ProductDetail";
-import LoginPage from "./pages/LoginPage";
-import Register from "./pages/Register";
+import LoginPage from "./pages/auth/LoginPage";
+import Register from "./pages/auth/Register";
 import PrivateRoute from "./components/PrivateRoute";
 import Dashboard from "./pages/admin/Dashboard";
 import AddProduct from "./pages/admin/AddProduct";
