+++ frontend/src/pages/auth/Register.jsx
